Rework: expose Gemma3 encoder surface instead of shipping allHiddenStates
Per review feedback, the encoder-style allHiddenStates extension was too
specialized for MLXLLM: its single uniform mask bypasses Gemma3's
per-layer sliding-window/global mask selection, which diverges from
generic Gemma3 semantics for texts longer than the sliding window. That
tap now lives in client code; this repo only exposes the minimal public
surface a client module needs to implement it:

- Gemma3TextConfiguration.hiddenSize, .hiddenLayers: internal -> public
- Gemma3TransformerBlock: internal -> public (init stays internal)
- Gemma3TransformerBlock.callAsFunction(_:mask:cache:): internal -> public
- Gemma3Model.embedTokens: internal -> public
- Gemma3Model.layers: internal -> public
- Gemma3Model.config: internal -> public

No behavior changes; access levels and doc comments only. `public` (not
`open`) since clients only call members from an extension; no
subclassing is required.

Gemma3EncoderAccessTests replaces the old test file: it implements the
same allHiddenStates tap inside the test using `import MLXLLM` without
`@testable`, proving the exposed surface is sufficient, and asserts
state count (numHiddenLayers + 1), (B, T, hiddenSize) shapes, first
state == scaled embedding, and determinism on a tiny
randomly-initialized model.
